Architecture!
Game is the global object that will collect all the globals we need for
the game and organize them nicely. Game specific methods should be added
to game.lua the same way Game:load_font is. We should do something
similar for other objects in the game, like the terrain representation,
and all that sort of thing.